refactor(학습로그 작성, 전체보기) : log -> post 패키지 이름 변경
